WL#13196 - XProtocol: Collection with json-schema validation must return error for concrete row
WL#12965 introduced a mechanism for validating data used in a collection.
Operations that are not matching with existing validation schema return
a generic error according to WL#12965 design. This error is just an indication
of a validation failure but do not give any information how the schema constraint
was violated.

This WL allows to get a more meaningful error for each operation on
a collection that violates the validation schema constraints. New error
describes how the schema was violated.

RB: 22363
Reviewed by: Grzegorz Szwarc <grzegorz.szwarc@oracle.com>
Reviewed by: Lukasz Kotula <lukasz.kotula@oracle.com>

Approved by: Shiva Shankar <shiva.y.shankar@oracle.com>