WL#10358: Implement table value constructors: VALUES
Implement the standard syntax for table value constructors (TVC).
The SQL standard lets us use VALUES syntax, i.e. TVC, wherever a SELECT
query specification may be used. This allows for statements such as
VALUES (1, 'one'), (2, 'two'), which previously only could be emulated
with UNIONS.

Note: The current MySQL server has a conflicting non-standard feature:
the VALUES() function of ON DUPLICATE KEY UPDATE update expressions,
which is a showstopper for the standard TVC syntax. However, the
standard also accepts a verbose form of TVC: VALUES ROW(), ROW() ..,
which is used instead.

Also extend <query specification> to allow an <explicit table> clause as
valid syntax. The syntax TABLE <table name> is equivalent to SELECT *
FROM <table name>.

The special syntax for INSERT statements, INSERT .. VALUES, remains as-is.
In addition, INSERT statements with VALUES that use the new verbose syntax
are rewritten to follow the same execution path as the old syntax.

Reviewed by: Roy Lyseng <roy.lyseng@oracle.com>