WL#10222 - MEB improvements for parallel DDL during backup
Introduce "single table locks". Before starting to copy a
tablespace, take a lock on one of the tables in the
tablespace, which allows DML, but no DDL on the tablespace.

The "single table locks" are implemented with
SQL statements:

    START TRANSACTION
    SELECT 1 FROM <table-id> LIMIT 1
    <copy the tablespace>
    COMMIT

The transaction must be kept open during the whole time of
the copy operation. Since multiple read threads might work
on multiple tablespaces in parallel, there must be one
connection available for each reader thread. These
connections will not be used for other purposes, so that the
transactions cannot be disturbed.

A consequence of the new "single table locks" is the need of
the privilege GRANT SELECT ON *.* TO <the backup user>.

In case of a general tablespace, one of the tables, using
it, is selected from information_schema.innodb_tables.

General tablespaces, that are "empty", that is, they don't
have any table, undo tablespaces, and fulltext tablespaces
are not locked.

Read space-id (and other attributes) from tablespaces after
having locked them to have correct knowledge about what is
being copied.

Add a fragment tracker to support multiple read threads.
A tablespace lock must be released only when all of its
fragments have been read.

Revise the final rescan algorithm, so that it does not only
[re-]copy tablespaces, but also renames or drops them, so
that the copied files list matches the current files list
during the backup lock.

Implement support for parallel RENAME TABLE and DROP TABLE.

When renaming table spaces in the rescan phase of a
directory backup, wait for the writers to be done with
the tablespaces before renaming them.

Extend the tablespace flag bits in the dictionary to
identify undo tablespaces. This allows undo tablespaces to
be treated correctly by the rescan algorithm.

Implement partial support for parallel DDL during an
optimistic backup. Retain the copied files list between the
two backup phases. Since these changes are incomplete for
now, detect optimistic backup (full and incremental) and
error out if such happens during an optimistic backup
operation.

Related changes:
================

Avoid exceptions on non-existent files when trying to
stat(2) them, as much as possible.

Modify some internal APIs so, that the size and time stamps
of non-existent files is returned as zero.

Modify code so that it can deal with tablespaces, that
vanish during a backup operation before they could be
locked.

For optimistic backup phase 1, do not try to open a file
when just its modification time is obtained.

For optimistic backup phase 1, do not rethrow an
Exception_file_not_found if a reader thread cannot open
a file.

In IO_file::get_file_time() do not write an error message,
if meb_stat() fails due to a missing file.

Modify checks in obtain_tablespace_id(), so that it can
handle short (or even empty) files gracefully. It can now
return SPACE_UNKNOWN. Check for it at certain places.

If a tablespace is too short for even the tablespace header
after locking it, skip the file.

Reuse read_tablespace_header() in TDE_util::is_table_tde_encrypted().

Fix --no-locking to have an effect on non-InnoDB tables only.
(That is, FLUSH TABLES <list> WITH READ LOCK.)
The new rescan algorithm does not work without a backup lock.

When validating pages be prepared to find encrypted pages even if
the tablespace header had not have the encryption flag. Mark the
tablespace as encrypted and stop verification on it.

Log InnoDB error messages as INFO. These messages are often
to be ignored by MEB. They shall not break the tests when
they pop up in test results.

Add injection point "6" after querying the server for InnoDB
files and before copying the first InnoDB file.

Add injection point "7" before each InnoDB file copy.

Add a new test suite meb_ddl.

Fix the existing test suite:
1. Accept result lines for the extra connection(s).
2. Firewall needs another recording for the select on a
   new table.
4. Accept warnings for empty general tablespaces.
5. Update MTR suppressions due to changed message format

RB#23022