Wl#12968: Configure replication applier to require row-based replication
This worklog aims to limit replication so that a channel only accepts
row-based replication.
This allows to restrict the type of instruction executed on the slave,
and as a side effect, limits the number of privileges a user needs when
associated to the replication applier.

Approved by: Pedro Figueiredo <pedro.figueiredo@oracle.com>
Approved by: Abhinav Agarwal <abhinav.ab.agarwal@oracle.com>