Bug#30108154 - SLOW STARTUP FOR MYSQL 8.0 WITH MANY TABLES DUE TO THE TABLESPACE FILES SCAN
The server startup is slow if there are large number of tables. This is because
all the tablespaces are scanned for space id while starting the server. This problem
is especially visible on slower disks.

This is because of the following two restrictions:
1) The multi-threaded scan kicks in only if the number of files is more than 50000
2) Three pages are scanned to get the space_id for each tablespace file

This fix changes the following:
1) Increased the parallelism by reducing the number of files scanned by each thread.
This effectively allocates more number of threads to scan the files
2) Read the space id from the first page only. This avoids multiple seeks needed to
read the space id from extra pages. If the space_id cannot be retrieved from the first
page, then the more heavyduty solution of getting the space id from all the pageѕ.

RB# 22877

Reviewed by: Kavin Lewis (kevin.lewis@oracle.com)
Reviewed by: Mayank Prasad (mayank.prasad@oracle.com)