Bug#30396902 stdx::expected<> move-assign fails with devstudio
Problem
=======

With DevStudio 12.6:

  stdx::expected<std::string, std::error_code> a("foo");

  a = stdx::expected<std::string, std::error_code>("bar");

results in 'a' containing "", instead of the expected "bar".

GCC, Clang, MSVC generate a implicit move-assign for stdx::expected
which:

  stdx::expected::operator=(expected &&other {
    stdx::ExpectedImpl::operator=(std::move(other));
    return *this;
  }

DevStudio 12.6 version is:

  stdx::expected::operator=(expected &&other) {
    stdx::ExpectedImpl::operator=(std::move(other));
    memmove(this, &other, sizeof(other));
    return *this;
  }

which leads to:

  other = std::string();

which then is memmove()d to *this again.

The same extra memmove() exists in the move-constructor.

Change
======

Instead of using an implicit move-assign and move-constructor, implement
it explicitly if devstudio is used.

additionally, explicitly allow default-construct non-copyable,
non-movable types

RB: 23139