Bug#30267889 ASSERTION `M_INDEX_CURSOR.IS_POSITIONED()' FAILED | TEMP (loosescan) [noclose]
NestedLoopSemiJoinWithDuplicateRemovalIterator takes a
single-table iterator as left argument, and inner-joins
it with the set of other SJ-inner tables. E.g. it doesn't work for
A SEMI JOIN (B LEFT JOIN C) with B as SJ-LooseScan table. So:
Disable SJ-loosescan if the 2nd table in the SJ nest is not
inner-joined to the first table.

Change-Id: Ic42070abda3b9c0ceade0c13cb9f76eed343031d