WL#13417 Handle the metadata upgrade in the Router WL#13188 Add support for Async Replica Sets in Router
Async Replica Sets are a new solution for creation and management of
asynchronous MySQL replication setups. They provide ease-of-use and
out-of-box usability similar to InnoDB Clusters for Group Replication setups.

WL#13417 adds to the Router handling of the upgrade procedure of the metadata
from version 1.x to 2.0. 2.0 is the version that introduces Async Replica Sets
and set of views on top of the schema tables that the metadata user is supposed
to use to access the metadata.

WL#13188 introduces support of the new metadata and Async Replica Sets in
the MySQL Router.

RB: 23043
RB: 22990