Bug#30382156: STORED PROCEDURE, JSON_TABLE AND "CONST" JOIN TYPE
If the first argument of JSON_TABLE was const during execution, but
not during preparation, it was not re-evaluated when a statement was
re-executed. The query returned an empty result on re-execution
because of this.

Fixed by resetting the flag that tells if the source has been
evaluated, at the end of execution.

Change-Id: Iab87e0769b586ad6afcaeb30f158c76caa87e572