Bug#28590016: INNODB TESTS FAIL ON PB2 WITH WARNINGS IN LOG
On startup, the log file contains a bunch of warnings that indicate the
"Doublewrite page ... could not be restored". for undo tablespace pages.
This happens if the engine aborts during undo tablespace truncation because
there are no MLOG_FILE_DELETE redo records for the deleted tablespace.
We can just skip these messages on undo space_ids that are being truncated.

Approved by Rahul Agarkar in RB#23126