Bug#30221187: MYSQLD FAILS WHEN COMMAND FROM PLUGIN EXECUTES TIMEOUT WITH LOG-SLOW-EXTRA
One of log-slow-extra's new log items is the last errno.
This is retrieved from the classic protocol object.
Unfortunately, that is the case even if that is not actually
the protocol in use.

Patch makes sure object is of the correct type before using it.
If a different protocol is used, errno information is not logged.

Approved by: Nisha Gopalakrishnan <nisha.gopalakrishnan@oracle.com>