BUG#30396844: Fix hang in LCP_WAIT_SYNC_EXTENT
When sync'ing extent pages we could end up stalling the LCP indefinitely
if a CONTINUEB of handling LCPs was outstanding when receiving the
FSWRITECONF of the last page written in the sync extent page. It was
a rare event and could also be restarted if another page was written
from the data pages. It is possible this mistake also caused
PREP_LCP pages to be written at times when they should not have been
written.

Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>