BUG#30305315 NDB_DD_BASIC CRASHES WHEN DROPPING LOGFILE GROUP
The next extent page to perform LCP on can be dropped during
drop of the logfile group file it belongs to. This can cause
the LCP to fail when accessing using the block variable
referring to this page. Fix is to ensure that the block variable
moves to the next page in the locked list if that happens.

Reviewed by: Mauritz Sundell <mauritz.sundell@oracle.com>