BUG#30273684: ASSERT IF GRANT IS EXECUTED ON RPL APPLIER AND APPLIER USER NOT WITH GRANT OPTIO
WL#12966: "Replication with privilege checks" enabled privilege
checks on replication channels, so that it is possible to replicate
securely from untrusted masters.
On the situation of the applier user not being able to grant
privileges, despite the applier was being stopped as expected,
additionally to that, a assert was triggered.

The above mentioned triggered was refactored to consider the
privilege checks introduced by WL#12966, that is, a scenario that
before was always successful now can be stopped by the lack of
privileges.

ReviewBoard: 23144