Bug#30358470 AUTO SYNC INCREASES LIKELIHOOD OF DEADLOCK WHILE ACQUIRING MDL
Problem:
--------

Auto sync can potentially trigger an increase in the number of locks being
taken on a particular metadata object at specific times such as when a
synchronization attempt coincides with a DDL statement or DML query involving
the same metadata object. The competing locks could lead to the deadlock
detection logic victimizing the user action rather than the background sync

Fix:
----

Change all exclusive metadata lock acquisition attempts during auto sync
to use a try lock strategy with a timeout of 0 in place of the erstwhile 10
seconds. This avoids deadlock detection and gives higher priority to the user
action

Change-Id: I2e6eb4d2f87a5a253a7c6ff68a6fe0bda3dd6acb