Bug#30402047 - WRONG USAGE OF GETADDRINFO RETURN VALUE
Description
===========
src/harness/src/networking/resolver.cc has the following line:
if (auto err = getaddrinfo(name, nullptr, &hints, &result) != 0) {

err type is deduced to be bool, but line below it is used like it was a
return value from getaddrinfo. So this line is missing parentheses or the
"!= 0" check is superfluous.

Fix
===
removed the "!= 0" check

RB: 23158