BUG#30406672: NDB VERSION NOT SET IN EARLY START PHASE FOR LGMAN
Reviewed-by: Frazer Clement <frazer.clement@oracle.com>

In 7.6 a new disk format was introduced. It was intended that one would be
able to start with the old format and still use that. However for other
reasons the upgrade to 7.6 required an initial node restart. This means
that 7.6 and onwards need not support the v1 format anymore.

There was startup code that didn't set the m_ndb_version until late in the
startup process. This led to initialisation of some pages being done according
to the old format and thus leading to crashes later on.

Since the old format is no longer required, the version is set to v2
immediately and if the files contains the old format we will crash since
it is no longer handled.