WL#9117: Improve NDB data node scalability
Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>

WL#9117: Decrease send_thread_mutex bottleneck
----------------------------------------------
The send_thread_mutex is a hot bottleneck when reaching towards full use of 16+ LDM threads
and when communication gets extensive. The solution lies in splitting this mutex such that
each send thread has its own mutex and that each block thread that assists with sending only
assists one send thread and not all of them. The send thread will each take care of a subset
of the transporters based on the id of the transporter.

Tried to also simplify handling of assist_send_thread. alert_send_thread will wake up all
send threads required except possibly the send thread the block thread is assisting. This
send thread will only be awakened if there is some transporters still needing sending after
assisting the send thread.

WL#9117: Decrease send buffer bottleneck
----------------------------------------
The send buffer mutex is another bottleneck that confronts the user of
NDB data nodes that have scalability issues. One problem in handling this
mutex is that the allocation of buffers happens in different threads
compared to where the memory is released.

The solution to this scalability issue is completely handled within mt.cpp
and its implementation of the thr_safe_pool. The thr_safe_pool actually
implements two different pools. The job buffer pool and the send buffer
pool. The main scalability concern is the send buffer pool and this is
what this patch targets to fix.

The solution lies in creating up to 16 different lists in the thr_safe_pool.
One list per send thread. The complexity comes from that release is not
balanced over the lists.

To handle scalability the aim is to try to only use one list. This list
is the list given by the send thread instance of the send thread or the
assisting send thread instance of the block thread.

For release_list we always use this list for small release amounts. For
larger amounts we loop over the lists and find the smallest list and
insert into this list. This looping is not using any locks and is thus
not secure to find the minimum. However the buffers in all lists are
available to anyone calling seize_list.

seize_list will try to allocate from the local list given by the above
send instance. If this list is empty it will perform a non-locking
search for the most filled instance. If this fails as well it will
scan through all lists to see if any has any free buffers available.

This means that allocation from the NDB Memory manager will only happen
when we have exhausted each and every list.

We added a mutex around allocation from NDB Memory manager to track the
amount of memory used totally in the send buffer pool and job buffer
pool.

Ensured that nosend threads are spread among the send thread instances as
well since they will still seize send buffers even if they will never
release them (since they never assist with sending).