WL#8426: Fully support 3 and 4 replicas (part)
Backport to 8.0.19

When using 3 replicas and 1 replica was the only alive node and the two
other nodes are restarting and crash in doing so, then also the live
node would crash. The problem is the function count_previously_alive_nodes
that counts nodes that only registered in QMGR as started. Actually it
turned out that no bitmask in any block have information about which
other nodes are started. To handle this a new bitmask c_cntr_startedNodeSet
was introduced in place of the old bitmask c_startedNodes. A new
bitmask c_startedNodeSet was introduced that contains a bit set for every
node that is fully started. This bitmask had to be sent in CNTR_START_CONF
message from master when accepting a node into the cluster.

By using this new c_startedNodeSet bitmask we can easily adapt
count_previously_alive_nodes to only count started nodes.

One more issue was that NDBCNTR allowed 2 nodes in the same node group
to simultaneously execute a node restart. The COPY_FRAGREQ implementation
was not designed for this. Rather than fixing this, we disallow multiple
nodes in the same node group from performing node restarts in parallel.
This has the advantage that we come back to a replicated state faster,
in particular when the copy fragment phase uses much resources from the
live node. It also avoids extra risk of overloading the live node.

Some more work to make test cases for multi sockets work properly with
3 and 4 replicas.

Reviewed-by: Frazer Clement <frazer.clement@oracle.com>