Bug#28765425 RPL_MULTI_SOURCE_CORRUPT_REPOSITORY FAILS AGAIN
Problem: Test 'rpl_multi_source_corrupt_repository' fails sporadically with
warning "file msr-relaylog.* was not purged because it was being read by thread
number *". An additional problem is that during test cleanup a binlog dump
thread is found in the list of running processes.

Analysis: Before processing some event, the replication applier is instructed to
move to the next relay-log and to purge the already applied logs. Method
log_in_use() is called and it detects that other 'THD' is accessing the relay
log. Therefore the applied relay logs are not purged and the warning is shown.

During cleanup the test calls 'rpl_create_msr_channels.inc' to recreate the
multi source replication channels. This script creates the first channel with an
empty name ' '. Then script 'rpl_end.inc' is called, which expects that the first
channel has name 'channel_1'. Due to mismatch in the name of the channel,
'rpl_end.inc' is not able to stop the slave for the first channel. Therefore a
binlog dump thread will be running on the master for that channel.

Solution: Suppress the warning "file * was not purged" since there is no
negative impact of it on either the normal operation of the server or on the
test scenario.

Do not call 'rpl_create_msr_channels.inc' to recreate the multi source
replication channels during test cleanup. These channels are stopped and deleted
at the end of include file "rpl_multi_source_corrupt_repository.inc". When
calling rpl_end.inc skip the steps for stopping the slave and deleting the
channels.

Add a new variable to 'rpl_end.inc' to specify if the step to stop the slave
should be skipped.

RB: 23176
Reviewed-by: Parveez Baig <parveez.baig@oracle.com>
Reviewed-by: Pedro Figueiredo <pedro.figueiredo@oracle.com>