Bug#30274618 - NDB : COMPLETED LCP WRONG DISPLAY FOR #BYTES
When a table fragment is checkpointed, BACKUP block sums up the
number of records and bytes of the fragment to the total statistics
checkpointed so far and sends the latter to the LQH Block.

Problem: When the LQH block receives it, it adds the received numbers
to its total statistics, and thus blowing up the statistics.

Solution : LQH block should reset its local statistics with the received
numbers, instead of adding them.

Reviewed-by: Priyanka Sangam <priyanka.sangam@oracle.com>