BUG#30343310 ASSERTION `BIT < MAP->N_BITS' POST SLAVE RESTART IF SLAVE HAS MORE COLUMNS
Description
-----------
An assertion is being hit while Applier privilege are being checked, after a
column is added on the slave.

Analysis
--------
In `Applier_security_context_guard::extract_columns_to_check` the set of
columns - for the table related to the event being processed - that need to be
checked for privileged access is composed. The loop that is used is testing for
the number of columns in the table definition instead of the number of columns
in the event. Therefore, if a column is added the number of columns in the
definition will exceed the number of columns in the event, leading to the
assertion.

Fix
---
Test for the number of columns in the event instead of the number of columns in
the table definition.

Reviewed-By: Pedro Gomes <pedro.gomes@oracle.com>
Reviewed-By: Luis Soares <luis.soares@oracle.com>
RB: 23110