Bug#30428543: PLUGINS USE "VESTIGIAL" WIRE_FORMAT_LITE_INL.H EVEN               WITH SYSTEM PROTOBUF > 3.8
The problem was that MySQL failed to compile with -DWITH_PROTOBUF=system
with protobuf 3.10. This was due to the inclusion of
google/protobuf/wire_format_lite_inl.h which was deprecated in 3.8
and removed in 3.10.

This patch fixes the problem by no longer including this header
as it is not needed.

Change-Id: I88c513713b96a1bf3b168da7fb585e7eaaaf5860