WL#13528: Remove integer display width from SHOW CREATE output
This patch changes SHOW CREATE and INFORMATION_SCHEMA output so that the
display width attribute is no longer included for integer columns without
ZEROFILL. Integer display width was deprecated in WL#13127.

This change means that we no longer generate CREATE syntax that we
have deprecated in cases where there is no behavior change.
Since ZEROFILL still is supported, integer display with is still
included for ZEROFILL columns.

The changes apply to tables, views and stored routines and to
SHOW CREATE, DESCRIBE statements and INFORMATION_SCHEMA queries.

Change-Id: Ib63cdde95647b31fb0ed96d00614b690f8652ca2