Bug #30439425 WL#8426 : VARIOUS TESTBLOBS AND TESTPARTITIONING TESTS FAILING WITH 3/4 REPLICAS
Problem:
--------

The maximum number of partitions that may be defined explicitly
for any NDB table is = 8 * [number of LDM threads] * [number of node groups]

setupUDPartitioning() function in testPartitioning ignores configurations with more than
2 replicas and thus the maximum number of partitions explicitly set exceeds the specification
which results in test failure.

Solution:
---------

Change the logic that calculates the number of partitions to be set on a
table explicitly to work for any number of replicas.

Approved-by: Mauritz Sundell <mauritz.sundell@oracle.com>