BUG#30226264 -- flush tables with read lock fails with deadlock found
FLUSH TABLES WITH READ LOCK causes a deadlock in case
LOCK INSTANCE FOR BACKUP statement was previously executed from within
the same session and there is an active ALTER DATABASE statement running
from another session against the same database as was specified (implicitly
or explicitly) for the FLUSH TABLES WITH READ LOCK statement.

The LOCK INSTANCE FOR BACKUP statement acquires BACKUP LOCK in exclusive mode.
During execution of the ALTER DATABASE statement mdl locks for namespaces
GLOBAL, SCHEMA and BACKUP are acquired exactly in the specified order. In result,
the ALTER DATABASE statement is blocked on waiting until BACKUP LOCK be released
while holding the MDL lock for database name specified in the ALTER DATABASE
statement.

Later when the statement FLUSH TABLES WITH READ LOCK is executed for some table
from the same database as specified in the ALTER DATABASE statement this table
being opened and the standalone function dd::mdl_lock_schema()
is called to acquire mdl lock for schema name that results in a deadlock.

To avoid deadlock for above use case the enumerator for BACKUP_LOCK
be moved upper in the enumeration enum_mdl_namespace. It leads to acquiring
mdl locks in the following order GLOBAL, BACKUP, SCHEMA during handling of
the ALTER DATABASE statement. In result deadlock is avoided for above mentioned
use case since an ALTER DATABASE statement would be blocked on BACKUP LOCK
without holding a lock on SCHEMA namespace.

Approved-by: Dyre Tjeldvoll <dyre.tjeldvoll@oracle.com>