Bug #30451479 : WL#8426 : "TESTNODERESTART -N GCPSTOPFALSEPOSITIVE T1"
Problem:
--------

A part of the test is to wait for the master node to reach the started state.
And the wait duration is proportional to the number of nodes in the cluster.
The wait duration is tuned for a 4 node cluster in the test case and thus
results in test failure for 6 node configurations.

Solution:
---------

Increase the wait duration to accommodate testing clusters with more
than 4 nodes as well.

Approved-by: Mauritz Sundell <mauritz.sundell@oracle.com>