Bug#30099539 - WL#13230: AUTOTEST TEST TESTNODERESTART -N PNR --NOLOGGING FAILURE
Problem: When 3 nodes recover in parallel in a 4-node 4-replica
constallation, one node has finished takeover and waiting for LCP to
complete, while another node is performing the takeover
currently. When the former node completes the LCP, the code releasing
it from the master takeover queue requires that this node must be the
current one performing takeover. This is a remnant from the 2-replica
constallation.

Fix: Relax the above requirement and accept it as a legal scenario.

Approved-by: Mikael Ronström <mikael.ronstrom@oracle.com>