Bug#30460528: RENAME FIELD::REAL_MAYBE_NULL() TO FIELD::IS_NULLABLE()
real_maybe_null() is a very confusing function name for a
function that returns if a Field is nullable or not.
These patches rename the function to make easier to understand
and to align it with e.g. Field::is_tmp_nullabe().

Part 1: Remove Field::maybe_null(). Replace it with calls to
Field::real_maybe_null() and (if needed) TABLE::is_nullable().

Change-Id: I13fa67500dd240e26a70c0c9806627f51458627a