Bug#30233405: SELF REFERENTIAL FKS ARE LOST WHEN THE TABLE IS ALTERED USING COPY ALGORITHM
When a table is altered using the copy algorithm, any self referential
foreign keys present on the table are lost. The ha_ndbcluster method
copy_fk_for_offline_alter() wrongly assumes that a foreign key on a
table can only refer columns from another table. Due to this, it fails
to properly copy the self referencing foreign keys after a copy alter.
This patch fixes that issue, reformats the method and adds a new
testcase to verify the fix. This patch also adds a few new tests to
ndb_ddl suite to verify the consistency of foreign keys after alter
DDLs.

Change-Id: Ia8763ac4b3b459e75c3f1a419da847f0f2aef267