BUG#30428328: GROUP_REPLICATION.GR_PRIMARY_MODE_GROUP_OPERATIONS_28 FAILS SPORADICALLY IN PB2
When a primary is elected in Group Replication single-primary mode,
it first updates its role and state, and only then unset
super_read_only, that is, a primary member can be seen as read only
for brief moments. Since we cannot make these three updates atomic,
this order: state, role, super_read_only; is the order that avoids
inconsistent status like a secondary member being writable.

To accommodate the above order, before we write to the primary
member we need to wait until it is writable.

ReviewBoard: 23236