Bug #30368622 MYSQL-8.0: NODES ARE CRASHING DURING RESTART WITH ERROR ""ANOTHER NODE FAILED DU
Improve robustness of Transporter / Protocol6 connection setup by having
clients send the nodeId of the server side that they think they are
connecting to.

This allows the server side to detect a mismatch here, and reject the
connection before it is established.

Prior to this change, the client was detecting the mismatch later, and
the resulting connection drop was treated as node failure, which could
cause a cascading failure of all starting nodes.

Logging of connection attempt failures is also modified so that
unexpected connection attempt failures are logged by data nodes, giving
better visibility of problems which may otherwise only be observed
as delays or issues with partial connectivity.

Upgrade/Downgrade unaffected by these changes as the new version server
copes with not receiving the server nodeid, and the old version server
copes with receiving extra unknown material on the HELLO line.

Approved by : Mikael Ronstrom <mikael.ronstrom@oracle.com>