Bug#29630735 - STACK BUFFER OVERFLOW IN LIBMYSQL (CLIENT.CC                FUNCTION READ_OK_EX) OUT OF BOUNDS WRITE
DESCRIPTION
===========
We encounter out of bounds write issue if the client
receives an incorrect session track in the OK packet.

FIX
===
Backported the fix from mysql-8.0 and made some changes
so that we have proper checks to validate the data.
If the packet turns out to be incorrect, an error is
returned.

RB: 23082