BUG#29205129 PFS.SOCKET_SUMMARY_* TABLES NOT UPDATED FOR SSL CONNECTIONS
Fix 1:

Prior to this fix, network io done at the socket level
was simply not instrumented with the performance schema
when using OpenSSL.

The root cause is lack of instrumentation in the OpenSSL code,
because OpenSSL is unaware of the performance schema wrappers api
such as mysql_socket_recv() to use in place of recv().

The fix is *not* to change the OpenSSL code,
but leverage the OpenSSL callbacks apis to add
extra actions before and after network io,
to add the missing performance schema instrumentation.

The OpenSSL api used is BIO_set_callback_ex().

With this change, network io performed by OpenSSL is
instrumented in the performance schema.

Fix 2:

Prior to this fix, when the server is IDLE waiting for the next command,
the socket used to read the command from was set in IDLE state.

In IDLE state, any network io performed was simply discarded.

The motivation for this was to avoid counting socket waits twice,
once as part of the idle wait and once as part of network io.

By dropping the event in the performance schema,
the number of bytes read and written was off.

In case of CLEAR communication, using a *byte* stream of data,
only the message length in a packet header was lost,
which is 4 bytes per query.

In case of SSL communications, using a *block* stream of data,
the protocol is different and decrypts an entire data block.

Not accounting the entire blocks of data in this case is a much
bigger flaw, so that the IDLE instrumentation has been changed
to better account the network io seen.

With this fix, a socket in IDLE state:
- does record an event, so that the number of bytes read and written
is accurate
- records a timing of 0 (considering the instrument un timed),
so that time spend is not counted twice with the IDLE instrument.

As a result:
- the 4 bytes offset error in CLEAR messages is fixed
- all network traffic is properly reported for SSL.

Approved by: Chris Powers <chris.powers@oracle.com>