Bug#27096081: MYSQLPUMP VALIDATES ALL DATABASE OBJECTS BY               IGNORING --INCLUDE_DATABASES OPTION
Description:- Mysqlpump utility fails to dump a database
when an invalid view object exsits in the system.

Analysis:- As per the current mysqlpump design, it first
searches and enumerates all the DB objects as part of the
crawler module and later filters out unwanted objects as
part of 'Chain Makers' module.
During the crawler stage, it enumerates all the DB objects
including views. While enumerating views, it triggers a
"LOCK TABLE <view> READ;" query to ensure all the view
dependent object exists. This query can return an error,
if one or more underlying base tables of the view are
dropped or if the user who defined this view is dropped.
If it returns an error, mysqlpump exits without being able
to dump. This happens even when the user wants to dump a
database which doesn't contain a view.

Fix:- View dependency check is moved post filteration.
This will ensure that only filtered out views will be
checked for their dependecies.

RB#23164