Bug#30484272 USING DISK DATA ON WINDOWS FAILS NODE RESTART: TSMAN: FILE SYSTEM CLOSE FAILED.
PROBLEM
=======

On Windows when restarting a non master data node with disk data it fails
with below log messages:

    2019-10-30 12:43:11 [ndbd] INFO     -- TSMAN: File system close failed.
OS errno: 6
    2019-10-30 12:43:11 [ndbd] ALERT    -- Node 1: Forced node shutdown
completed. Occurred during startphase 0.

This is caused by a bug in Win32AsyncFile::openReq() there the error
checking for a call to GetFileSizeEx() was reversed.

    void Win32AsyncFile::openReq(Request* request)
    ...,
      if (flags & FsOpenReq::OM_READ_SIZE)
      {
        LARGE_INTEGER size;
        BOOL ret_code = GetFileSizeEx(hFile, &size);
        if (ret_code) <<<<--- ERROR should be: if (!ret_code)
        {
          request->error = GetLastError();
          (void)CloseHandle(hFile);

SOLUTION
========

Treat false from GetFileSizeEx() as error.

Reviewed-by: Frazer Clement <frazer.clement@oracle.com>