Bug #30412632: JSON SCHEMA DOESN'T RESPOND CORRECTLY IN CERTAIN SITUATIONS
Description:
We have a few scenarios that don't seem to behave correctly as expected.
All the scenarios involve the options passed to it.
1.  It is our belief that passing { validation: {} } to create
collection should work and leave you with default settings.
Currently it throws invalid # of arguements.
2.  Passing { validation: { level: "off" }} to create collection
should also work since that is a default state. Currently it throws
invalid # of arguments
3.  Passing { validation: { foo: "bar" }} should throw an error about an
unknown keyword 'foo'. Right now it throws invalid # of arguments.
It seems as this all stems from the requirement that the schema property
be included with create collection.  #1 and #2 is a matter of opinion
and reasonable people can agree they are not bugs.

RB:23277
Reviewed-by: Lukasz Kotula <lukasz.kotula@oracle.com>
Change-Id: I3528dc1dbcc8da97e74341908f999e5d24706396