Bug#30482553: CMAKE SKIPS BUILDING WITH SASL ON SOLARIS 11.4 DUE TO BAD SASL.H
The problem was that Cyrus SASL 2.1.26 which is used by Solaris 11.4 has
a bug where sasl/sasl.h is not self contained but requires sys/types.h
to be included first. This caused CMake checks to fail even if they
should not. Which in turn made it impossible to build the LDAP
authentication plugin.

This patch fixes the problem by including sys/types.h before
doing the CMake check for the existence of sasl/sasl.h.

Change-Id: Ib6448e9c030edb5e7dd1968a1281d2ac5561f016